requirements-full.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.7
botocore==1.38.7
Brotli==1.1.0
browser-use==0.1.41
browserforge==1.2.3
cachetools==5.5.2
camoufox==0.4.11
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
configuraptor==1.27.2
contourpy==1.3.2
coverage==7.8.0
cssselect==1.3.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
faiss-cpu==1.11.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
geoip2==5.1.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.0.10
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-openrouter==0.0.1
langchain-text-splitters==0.3.7
langsmith==0.3.39
language-tags==1.2.0
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.3
maxminddb==2.7.0
mcp==1.6.0
mcp-python==0.1.4
mcp-server-git==2025.1.14
mdurl==0.1.2
mem0ai==0.1.88
monotonic==1.6
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.3
networkx==3.4.2
numpy==1.26.4
ollama==0.4.8
openai==1.76.2
orjson==3.10.18
packaging==24.2
pandas==2.2.3
patchright==1.51.3
pillow==11.2.1
platformdirs==4.3.8
playwright==1.51.0
pluggy==1.6.0
plumbum==1.9.0
portalocker==2.10.1
posthog==3.25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyee==12.0.0
Pygments==2.19.1
pyodbc==5.2.0
pyparsing==3.2.3
pyperclip==1.9.0
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
rebrowser_playwright==1.49.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
result==0.17.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scrapling==0.2.99
scrapling-fetch-mcp==0.1.6
screeninfo==0.8.1
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.3.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadful==0.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.1
tomli==2.2.1
tomli_w==1.2.0
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typeguard==4.4.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.4.0
uv==0.7.2
uvicorn==0.34.2
uvx==1.0.3
w3lib==2.3.1
yarl==1.20.0
zstandard==0.23.0
